Calc2KeyCE.Universal/requirements.txt
altgraph==0.17.2
cffi==1.15.0
click==8.0.3
clr-loader==0.1.6
colorama==0.4.4
EasyProcess==0.3
entrypoint2==0.2.4
future==0.18.2
importlib-metadata==4.8.2
mss==6.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
pefile==2021.9.3
Pillow==8.4.0
pip==21.3.1
platformdirs==2.4.0
pycparser==2.21
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
pyusb==1.2.1
pywin32-ctypes==0.2.0
regex==2021.11.10
setuptools==59.2.0
tomli==1.2.2
typed-ast==1.5.0
typing_extensions==4.0.0
Wand==0.6.7
zipp==3.6.0
git+git://github.com/pythonnet/pythonnet.git@master#egg=pythonnet